Refactor publish result and preflight handling (#59)
Extract cargo publish result handling into a dedicated helper so the
publish command path delegates success, already-published, index-miss,
and failure outcomes through one named result handler.

Move publish pre-flight helpers and publish error types into focused
modules, remove the file-level module-size suppression from `publish.py`,
and keep compatibility aliases for existing tests and callers.
