+++ lading/commands/publish.py
@@ -1,21 +1,18 @@
 """Publication planning helpers for :mod:`lading.commands.publish`."""
 
-# pylint: disable=too-many-lines
-
 from __future__ import annotations
 
 import atexit
-import collections.abc as cabc
 import dataclasses as dc
 import logging
-import os
 import shutil
 import tempfile
 import typing as typ
 from pathlib import Path
 
 from lading import config as config_module
-from lading.commands.publish_diagnostics import _append_compiletest_diagnostics
+from lading.commands import publish_preflight as _publish_preflight
+from lading.commands.publish_errors import PublishError, PublishPreflightError
 from lading.commands.publish_execution import (
     _CommandRunner,
     _invoke,
@@ -58,23 +55,21 @@ _should_use_cmd_mox_stub = should_use_cmd_mox_stub
 _split_command = split_command
 _append_section = append_section
 _format_plan = format_plan
+_CargoPreflightOptions = _publish_preflight._CargoPreflightOptions
+_apply_compiletest_externs = _publish_preflight._apply_compiletest_externs
+_build_preflight_environment = _publish_preflight._build_preflight_environment
+_build_test_arguments = _publish_preflight._build_test_arguments
+_compose_preflight_arguments = _publish_preflight._compose_preflight_arguments
+_normalise_test_excludes = _publish_preflight._normalise_test_excludes
+_run_aux_build_commands = _publish_preflight._run_aux_build_commands
+_run_cargo_preflight = _publish_preflight._run_cargo_preflight
+_verify_clean_working_tree = _publish_preflight._verify_clean_working_tree
 
 LOGGER = logging.getLogger(__name__)
 
 if typ.TYPE_CHECKING:
     from lading.config import LadingConfig
     from lading.workspace import WorkspaceCrate, WorkspaceGraph
-
-
-@dc.dataclass(frozen=True, slots=True)
-class _CargoPreflightOptions:
-    """Options controlling how cargo pre-flight commands are invoked."""
-
-    extra_args: cabc.Sequence[str]
-    test_excludes: cabc.Sequence[str] = ()
-    unit_tests_only: bool = False
-    env: cabc.Mapping[str, str] | None = None
-    diagnostics_tail_lines: int | None = None
 
 
 @dc.dataclass(frozen=True, slots=True)
@@ -151,73 +146,6 @@ class PublishPreparation:
 
     staging_root: Path
     copied_readmes: tuple[Path, ...]
-
-
-class PublishPreflightError(RuntimeError):
-    """Raised when required pre-publication checks fail."""
-
-
-class PublishError(PublishPreflightError):
-    """Raised when publishing crates fails after pre-flight checks."""
-
-
-def _build_preflight_environment(
-    overrides: tuple[tuple[str, str], ...],
-) -> dict[str, str]:
-    """Return the base environment for publish pre-flight commands."""
-    env = dict(os.environ)
-    env.update(overrides)
-    return env
-
-
-def _run_aux_build_commands(
-    workspace_root: Path,
-    commands: tuple[tuple[str, ...], ...],
-    *,
-    runner: _CommandRunner,
-    env: cabc.Mapping[str, str] | None,
-) -> None:
-    """Execute auxiliary build commands prior to cargo pre-flight runs."""
-    for command in commands:
-        exit_code, stdout, stderr = runner(command, cwd=workspace_root, env=env)
-        if exit_code != 0:
-            detail = (stderr or stdout).strip()
-            rendered = " ".join(command)
-            message = (
-                f"Auxiliary build command failed with exit code {exit_code}: {rendered}"
-            )
-            if detail:
-                message = f"{message}; {detail}"
-            raise PublishPreflightError(message)
-
-
-def _resolve_extern_path(workspace_root: Path, raw_path: str) -> Path:
-    """Return ``raw_path`` resolved relative to ``workspace_root`` when needed."""
-    candidate = Path(raw_path)
-    if not candidate.is_absolute():
-        candidate = workspace_root / candidate
-    return candidate.expanduser().resolve(strict=False)
-
-
-def _apply_compiletest_externs(
-    env: cabc.Mapping[str, str],
-    externs: tuple[config_module.CompiletestExtern, ...],
-    *,
-    workspace_root: Path,
-) -> dict[str, str]:
-    """Return ``env`` with compiletest externs appended to ``RUSTFLAGS``."""
-    if not externs:
-        return dict(env)
-    updated = dict(env)
-    flags = " ".join(
-        f"--extern {extern.crate}={_resolve_extern_path(workspace_root, extern.path)}"
-        for extern in externs
-    ).strip()
-    if not flags:
-        return updated
-    previous = updated.get("RUSTFLAGS", "").strip()
-    updated["RUSTFLAGS"] = " ".join(filter(None, (previous, flags)))
-    return updated
 
 
 def _normalise_build_directory(
@@ -388,10 +316,8 @@ def _handle_index_missing_version(
 ) -> None:
     """Pick the phase-appropriate error class and delegate to the helper.
 
-    ``publish_index_check`` cannot import :class:`PublishError` /
-    :class:`PublishPreflightError` without a circular dependency, so the
-    error class is resolved here based on the cargo subcommand and passed
-    through to the relocated implementation.
+    Resolve the error class here based on the cargo subcommand and pass it
+    through to the relocated implementation in ``publish_index_check``.
     """
     error_cls = (
         PublishError if invocation.subcommand == "publish" else PublishPreflightError
@@ -514,6 +440,26 @@ def _publish_crate(
         cwd=crate_root,
         env=None,
     )
+    _handle_publish_result(
+        _CargoInvocation(
+            crate_name=crate.name,
+            subcommand="publish",
+            output=(exit_code, stdout, stderr),
+        ),
+        crate,
+        plan,
+        options,
+    )
+
+
+def _handle_publish_result(
+    invocation: _CargoInvocation,
+    crate: WorkspaceCrate,
+    plan: PublishPlan,
+    options: _PublishExecutionOptions,
+) -> None:
+    """Handle a completed ``cargo publish`` invocation."""
+    exit_code, stdout, stderr = invocation.output
     if exit_code == 0:
         return
     if _is_already_published_error(exit_code, stdout, stderr):
@@ -527,15 +473,7 @@ def _publish_crate(
         # cargo publish --dry-run packages internally and hits the same
         # crates.io index lookup as cargo package, so honour the override
         # consistently across both phases.
-        _handle_index_missing_version(
-            _CargoInvocation(
-                crate_name=crate.name,
-                subcommand="publish",
-                output=(exit_code, stdout, stderr),
-            ),
-            plan=plan,
-            options=options,
-        )
+        _handle_index_missing_version(invocation, plan=plan, options=options)
         return
 
     message = _format_cargo_failure_message(
@@ -743,17 +681,6 @@ def _run_preflight_checks(
         )
 
 
-def _compose_preflight_arguments(
-    target_dir: Path, *, include_all_targets: bool
-) -> tuple[str, ...]:
-    """Build the ordered argument tuple shared by pre-flight cargo commands."""
-    arguments = ["--workspace"]
-    if include_all_targets:
-        arguments.append("--all-targets")
-    arguments.append(f"--target-dir={target_dir}")
-    return tuple(arguments)
-
-
 def _preflight_argument_sets(
     target_dir: Path, *, unit_tests_only: bool
 ) -> tuple[tuple[str, ...], tuple[str, ...]]:
@@ -763,93 +690,3 @@ def _preflight_argument_sets(
         target_dir, include_all_targets=not unit_tests_only
     )
     return check_arguments, test_arguments
-
-
-def _normalise_test_excludes(entries: cabc.Sequence[str]) -> tuple[str, ...]:
-    """Return sorted, deduplicated, trimmed crate names for ``--exclude`` flags."""
-    return tuple(sorted({crate.strip() for crate in entries if crate.strip()}))
-
-
-def _build_test_arguments(
-    base_arguments: list[str], options: _CargoPreflightOptions
-) -> list[str]:
-    """Return cargo test arguments derived from ``options``."""
-    arguments = list(base_arguments)
-    if options.unit_tests_only:
-        arguments.extend(("--lib", "--bins"))
-    for crate_name in _normalise_test_excludes(options.test_excludes):
-        # Sorted unique values keep cargo invocations deterministic for tests/logging.
-        arguments.extend(("--exclude", crate_name))
-    return arguments
-
-
-def _verify_clean_working_tree(
-    workspace_root: Path,
-    *,
-    allow_dirty: bool,
-    runner: _CommandRunner,
-    env: cabc.Mapping[str, str] | None = None,
-) -> None:
-    """Ensure ``workspace_root`` has no uncommitted changes unless allowed."""
-    if allow_dirty:
-        return
-
-    exit_code, stdout, stderr = runner(
-        ("git", "status", "--porcelain"),
-        cwd=workspace_root,
-        env=env,
-    )
-    if exit_code != 0:
-        detail = (stderr or stdout).strip()
-        message = (
-            "Failed to verify workspace state; is this a git repository?"
-            if "not a git repository" in detail.lower()
-            else "Failed to verify workspace state with git status"
-        )
-        if detail:
-            message = f"{message}: {detail}"
-        raise PublishPreflightError(message)
-    if stdout.strip():
-        message = (
-            "Workspace has uncommitted changes; commit or stash them "
-            "before publishing or re-run without --forbid-dirty."
-        )
-        raise PublishPreflightError(message)
-
-
-def _run_cargo_preflight(
-    workspace_root: Path,
-    subcommand: typ.Literal["check", "test"],
-    *,
-    runner: _CommandRunner,
-    options: _CargoPreflightOptions,
-) -> None:
-    """Run ``cargo <subcommand>`` inside ``workspace_root``."""
-    arguments = list(options.extra_args)
-    if subcommand == "test":
-        arguments = _build_test_arguments(arguments, options)
-    exit_code, stdout, stderr = runner(
-        ("cargo", subcommand, *arguments),
-        cwd=workspace_root,
-        env=options.env,
-    )
-    if exit_code != 0:
-        message = _build_cargo_error_message(subcommand, exit_code, stdout, stderr)
-        if options.diagnostics_tail_lines is not None:
-            message = _append_compiletest_diagnostics(
-                message,
-                stdout,
-                stderr,
-                tail_lines=options.diagnostics_tail_lines,
-            )
-        raise PublishPreflightError(message)
-
-
-def _build_cargo_error_message(
-    subcommand: str, exit_code: int, stdout: str, stderr: str
-) -> str:
-    """Return a consistent failure message for cargo pre-flight commands."""
-    message = f"Pre-flight cargo {subcommand} failed with exit code {exit_code}"
-    if detail := (stderr or stdout).strip():
-        message = f"{message}: {detail}"
-    return message
